% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SangerVariants.R
\name{callSangerVariants_fasta}
\alias{callSangerVariants_fasta}
\title{Title}
\usage{
callSangerVariants_fasta(
  sampleKey_file,
  fasta_path,
  fasta_extension = ".fa",
  reference_files,
  hardTrim = c(0, 0)
)
}
\arguments{
\item{sampleKey_file}{File name to a table specifying the samples.
This table must have the following columns:
* SampleID: a unique ID for each bacterial sample
* Species
* Strain
* Gene
* Primer: the name of the primer used for the sequencing
* FileName: the name of the fasta file (without file extension)
* Success: whether or not the sequencing was successful (TRUE or FALSE)
* Reference: the name of the corresponding reference sequence}

\item{fasta_path}{The path where the fasta files are located.}

\item{fasta_extension}{File extension of the fasta files, usually either ".fa" (default) or ".fasta".}

\item{reference_files}{The path and file name(s) of the fasta file(s) with the reference sequences}

\item{hardTrim}{A vector specifying how many nucleotides on each side of a sequence should be hard-trimmed.}
}
\value{
A list containing two tables: "summary" list the number of variants of different types detected for each sample,
whereas "details" provides detailed information for each variant.
}
\description{
Title
}
