% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutTable.R
\name{fillMutationsTable}
\alias{fillMutationsTable}
\title{Fill in missing data in a mutations table.}
\usage{
fillMutationsTable(muts, refs, seqs, coordinates)
}
\arguments{
\item{muts}{An (incomplete) mutations table.}

\item{refs}{A table specifying the species and names for all reference sequences.}

\item{seqs}{Reference sequences as a DNAStringSet object, with names as specified in refs.}

\item{coordinates}{A list of coordinates, with names corresponding to seqs.}
}
\value{
A muts table that is (hopefully) more complete than the input table.
}
\description{
This function tries to infer missing data in a mutations table,
including mutations at the amino acid level when only the nucleotide
level information is present, explicit positions of mutations where
this data is only implicitly available, and positions in reference (E.coli)
coordinates.
The result is a standardised table that can then more readily be analysed.
}
