% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{getAllCoordinates}
\alias{getAllCoordinates}
\title{Obtain tables of coordinates.}
\usage{
getAllCoordinates(seqs, refSeq_ID)
}
\arguments{
\item{seqs}{A DNAStringSet object containing a set of homologous sequences.}

\item{refSeq_ID}{The name of the sequence designated as the reference.}
}
\value{
A list of tables of coordinates. Each of these tables has two columns:
the position in the focal sequence (posFocal), and the position in the reference sequence (posRef).
}
\description{
Takes a list of homologous sequences of which one is designated a reference,
aligns them all to the reference (at amino-acid level), and calculates for each sequence
a table of corresponding coordinates.
}
