% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutTable.R
\name{isPointMutation}
\alias{isPointMutation}
\title{Checks if a row in a mutations table indicates a point mutation}
\usage{
isPointMutation(muts, k)
}
\arguments{
\item{muts}{A mutations table, as generated by `fillMutationsTable`.}

\item{k}{Row within that table.}
}
\value{
TRUE or FALSE
}
\description{
Checks if a row in a mutations table indicates a point mutation
}
\keyword{internal}
