% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutTable.R
\name{mutationsTableSummary}
\alias{mutationsTableSummary}
\title{Print a summary of a mutations table}
\usage{
mutationsTableSummary(muts)
}
\arguments{
\item{muts}{A mutations table, as created using the fillMutationsTable function.}
}
\description{
Print a summary of a mutations table
}
