% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{translateCoordinate}
\alias{translateCoordinate}
\title{Translates between positions in two sequences.}
\usage{
translateCoordinate(
  pos,
  coordinates,
  direction = "FocalToRef",
  AAinput = FALSE,
  AAoutput = FALSE
)
}
\arguments{
\item{pos}{Position within the sequence.}

\item{coordinates}{Table of coordinates.}

\item{direction}{Either "FocalToRef" (the default) or "RefToFocal".}

\item{AAinput}{When `FALSE` (the default), the `pos` argument will be assumed to be a
nucleotide position. Otherwise, `pos` will be treated as an amino acid position.}

\item{AAoutput}{When `FALSE` (the default), the returned translated position will be a nucleotide
position, otherwise an amino acid position.}
}
\value{
The translated position.
}
\description{
Using coordinates calculated using the `getCoordinates` or `getAllCoordinates` function,
this function translates between a position in a focal sequence to a position in
the reference sequence, or vice versa.
}
