% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{getCoordinates}
\alias{getCoordinates}
\title{Calculate a table of coordinates}
\usage{
getCoordinates(dnaFocal, dnaRef)
}
\arguments{
\item{dnaFocal}{Focal DNA sequence.}

\item{dnaRef}{Reference DNA sequence, e.g. Escherichia coli.}
}
\value{
A table with two columns: the position in the focal sequence (posFocal),
and the position in the reference sequence (posRef).
}
\description{
This function takes two DNA sequences, aligns the translated amino acid sequences,
and calculates a table of corresponding positions in this alignment ("coordinates").
}
